pyproject.toml
[tool.poetry]
name = "notion-autotests-mobile"
version = "0.1.0"
description = ""
authors = ["Sergey Galtsov <galtsovsergey@mail.ru>"]
readme = "README.md"
packages = [{include = "notion_autotests_mobile"}]

[tool.poetry.dependencies]
python = "^3.10"
selene = {version = "^2.0.0b17", allow-prereleases = true}
appium-python-client = "^2.8.1"
pytest = "^7.2.2"
pytest-xdist = "^3.2.1"
allure-pytest = "^2.13.1"
requests = "^2.28.2"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
